# Add score_threshold and max_num to NMS